# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Agents - HTTP API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.27.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/sovereignty/access/check": {
      "get": {
        "tags": [
          "sovereignty"
        ],
        "operationId": "sovereignty_check_access",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Data category to check",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing category parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sovereignty/consent/grant": {
      "post": {
        "tags": [
          "sovereignty"
        ],
        "operationId": "sovereignty_grant_consent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SovereigntyConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Consent granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SovereigntyConsentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sovereignty/consent/revoke": {
      "post": {
        "tags": [
          "sovereignty"
        ],
        "operationId": "sovereignty_revoke_consent",
        "responses": {
          "200": {
            "description": "Consent revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SovereigntyConsentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Consent not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sovereignty/status": {
      "get": {
        "tags": [
          "sovereignty"
        ],
        "operationId": "sovereignty_status",
        "responses": {
          "200": {
            "description": "Sovereignty status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SovereigntyStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "List templates",
        "operationId": "list_templates",
        "responses": {
          "200": {
            "description": "List of templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/templates/{id}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Get template by ID",
        "operationId": "get_template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessCheckResponse": {
        "type": "object",
        "required": [
          "category",
          "classification",
          "access_required"
        ],
        "properties": {
          "access_required": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          }
        }
      },
...
      "SovereigntyConsentRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SovereigntyConsentResponse": {
        "type": "object",
        "required": [
          "consent",
          "message",
          "categories"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "consent": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SovereigntyStatusResponse": {
        "type": "object",
        "required": [
          "explicit_consent",
          "requires_affirmative_consent",
          "sovereign_data",
          "shared_data",
          "public_data",
          "granted_categories"
        ],
        "properties": {
          "explicit_consent": {
            "type": "boolean"
          },
          "granted_categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "public_data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requires_affirmative_consent": {
            "type": "string"
          },
          "shared_data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sovereign_data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TemplateResponse": {
        "type": "object",
        "description": "Template response",
        "required": [
          "id",
          "template_type",
          "name",
          "description",
          "source_path",
          "lexicon_terms"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lexicon_terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "source_path": {
            "type": "string"
          },
          "template_type": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}